rename: rename to noesy_neighbors
