+++ smoltools/__init__.py
@@ -1,3 +1,3 @@
 import smoltools.pdbtools as pdbtools
 import smoltools.fret0 as fret0
-import smoltools.albatrosy as albatrosy
+import smoltools.noesy_neighbors as noesy_neighbors
